feat(ui): Add data migration files and enhance CSV handling in order and chart modules
