+++ common/changes/@yuants/data-interest-rate/2025-08-16-08-55.json
@@ -1,0 +1,10 @@
+{
+  "changes": [
+    {
+      "packageName": "@yuants/data-interest-rate",
+      "comment": "fix migration",
+      "type": "patch"
+    }
+  ],
+  "packageName": "@yuants/data-interest-rate"
+}
