+++ common/changes/@yuants/data-product/2025-08-16-08-55.json
@@ -1,0 +1,10 @@
+{
+  "changes": [
+    {
+      "packageName": "@yuants/data-product",
+      "comment": "fix migration",
+      "type": "patch"
+    }
+  ],
+  "packageName": "@yuants/data-product"
+}
