+++ common/changes/@yuants/sql/2025-08-16-08-55.json
@@ -1,0 +1,10 @@
+{
+  "changes": [
+    {
+      "packageName": "@yuants/sql",
+      "comment": "fix migration",
+      "type": "patch"
+    }
+  ],
+  "packageName": "@yuants/sql"
+}
