+++ libraries/data-order/src/migration.ts
@@ -47,6 +47,6 @@ AddMigration({
   name: 'add_column_order_traded_value',
   dependencies: ['dcc5ee42-00da-4eca-8abc-6bb165dcdd41'],
   statement: `
-    ALTER TABLE "order" ADD COLUMN IF NOT EXISTS traded_value DECIMAL(20, 8)
+    ALTER TABLE "order" ADD COLUMN IF NOT EXISTS traded_value DECIMAL(20, 8);
   `,
 });
