+++ ui/web/package.json
@@ -74,7 +74,9 @@
     "ethers": "~6.13.5",
     "@dnd-kit/core": "~6.3.1",
     "@dnd-kit/sortable": "~10.0.0",
-    "@dnd-kit/utilities": "~3.2.2"
+    "@dnd-kit/utilities": "~3.2.2",
+    "@yuants/data-quote": "workspace:*",
+    "@yuants/data-interest-rate": "workspace:*"
   },
   "devDependencies": {
     "@originjs/vite-plugin-commonjs": "~1.0.3",
