+++ ui/web/src/modules/Order/OrderListPanel.tsx
@@ -1,12 +1,12 @@
 import { IconExport } from '@douyinfe/semi-icons';
 import { Button, Space } from '@douyinfe/semi-ui';
 import { formatTime } from '@yuants/utils';
-import { stringify } from 'csv-stringify/browser/esm/sync';
 import download from 'downloadjs';
 import { useObservableState } from 'observable-hooks';
 import { useTranslation } from 'react-i18next';
 import { DataView } from '../Interactive';
 import { registerPage } from '../Pages';
+import { CSV } from '../Util';
 import { orders$ } from './model';
 
 registerPage('OrderListPanel', () => {
@@ -14,27 +14,7 @@ registerPage('OrderListPanel', () => {
   const orders = useObservableState(orders$);
 
   const handleExportOrderList = () => {
-    download(
-      stringify(orders, {
-        header: true,
-        columns: [
-          'order_id',
-          'product_id',
-          'submit_at',
-          'order_direction',
-          'order_type',
-          'traded_price',
-          'traded_volume',
-          'traded_value',
-          'price',
-          'volume',
-          'profit_correction',
-          'comment',
-        ],
-      }),
-      'orders.csv',
-      'text/csv',
-    );
+    download(CSV.stringify(orders), 'orders.csv', 'text/csv');
   };
 
   return (
